% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cancel.fraction}
\alias{cancel.fraction}
\title{cancel a fraction to the smallest numbers}
\usage{
cancel.fraction(numerator, denominator)
}
\arguments{
\item{numerator}{the fraction's numerator}

\item{denominator}{the fraction's denominator
#'}
}
\value{
An integer vector of length 2 containing the reduced numerator and denominator.
}
\description{
cancel a fraction to the smallest numbers
}
\examples{
cancel.fraction(40,15)
cancel.fraction(42, 56)

}
