\name{gcd}
\alias{gcd}

\title{ Greatest common divisor of two numbers }
\description{
  This function gives the greatest common divisor of two numbers
}
\usage{
gcd(x, y)
}
\arguments{
  \item{x}{ first number }
  \item{y}{ second number }
}


\author{ Joerg Schlarmann }

\seealso{  \code{\link{scm}} }
\examples{
gcd(42, 56)
}

\keyword{ math }

