% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratics.R
\name{getfactorisingproblem}
\alias{getfactorisingproblem}
\title{getfactorisingproblem()}
\usage{
getfactorisingproblem()
}
\value{
List containing Coefficients, String Representation and Solution
}
\description{
Get Problem to Factorise
}
\examples{
x = getfactorisingproblem()
cat(x[[2]])
}
\author{
Chris Casey
}
