% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratics.R
\name{getproblems}
\alias{getproblems}
\title{getproblems(n)}
\usage{
getproblems(n)
}
\arguments{
\item{n}{Number of problems}
}
\value{
None - printed to screen
}
\description{
Get list of factorising problems
}
\examples{
getproblems(10)

}
