% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is.decimal}
\alias{is.decimal}
\title{checks if a number is decimal or integer}
\usage{
is.decimal(x)
}
\arguments{
\item{x}{the number to check
#'}
}
\value{
true or false
}
\description{
checks if a number is decimal or integer
}
\examples{
is.decimal(40.15)
is.decimal(4015)

}
