% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is.even}
\alias{is.even}
\title{checks if a number or vector is even}
\usage{
is.even(x)
}
\arguments{
\item{x}{the number to check
#'}
}
\value{
true or false
}
\description{
checks if a number or vector is even
}
\examples{
is.even(45)
is.even(46)
x <- c(1,2,3,4,5, 6, 7)
is.even(x)

}
