% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is.negative}
\alias{is.negative}
\title{check whether numbers of a vector are negative}
\usage{
is.negative(x)
}
\arguments{
\item{x}{the number or vector to check
#'}
}
\value{
true or false
}
\description{
check whether numbers of a vector are negative
}
\examples{
is.negative(3)  # this will return FALSE
is.negative(-2)  # this will return TRUE

x <- c(-1, -2, 3.02, 4, -5.2, 6, -7)
is.negative(x)

}
