% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is.positive}
\alias{is.positive}
\title{check whether numbers of a vector are positive}
\usage{
is.positive(x)
}
\arguments{
\item{x}{the number or vector to check
#'}
}
\value{
true or false
}
\description{
check whether numbers of a vector are positive
}
\examples{
is.positive(-3)   # this will return FALSE
is.positive(2)  # this will return TRUE

x <- c(-1, -2, 3.02, 4, -5.2, 6, -7)
is.positive(x)

}
