% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is.prim}
\alias{is.prim}
\title{Check whether a vector contains prime numbers}
\usage{
is.prim(x)
}
\arguments{
\item{x}{The number or vector to check.}
}
\value{
A logical vector indicating whether values are prime.
}
\description{
Check whether a vector contains prime numbers
}
\examples{
is.prim(8)
is.prim(11)

x <- c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11)
is.prim(x)

}
