% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is.real.positive}
\alias{is.real.positive}
\title{check whether numbers of a vector are real positive. Real positive means, that zero is included as a positive number.}
\usage{
is.real.positive(x)
}
\arguments{
\item{x}{the number or vector to check
#'}
}
\value{
true or false
}
\description{
check whether numbers of a vector are real positive. Real positive means, that zero is included as a positive number.
}
\examples{
is.real.positive(-3)   # this will return FALSE
is.real.positive(0)    # this will return TRUE

x <- c(0, -1, -2, 3.02, 4, -5.2, 6, -7)
is.real.positive(x)

}
