\name{prime.factor}
\alias{prime.factor}
\title{ giving prime-factors of a number }
\description{
  This function calculates the prime-factors of a number
  }
\usage{
prime.factor(n)
}
\arguments{
  \item{n}{ the number to be checked }
}

\author{ Joerg Schlarmann }

\examples{
prime.factor(21)
prime.factor(100)
}

\keyword{ math }

