% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{primes}
\alias{primes}
\title{generate prime-numbers in a range from START to END}
\usage{
primes(start = 12, end = 9999)
}
\arguments{
\item{start}{the number to start from}

\item{end}{the number to end
#'}
}
\value{
a vector of prime numbers
}
\description{
generate prime-numbers in a range from START to END
}
\examples{
primes(12,150) # list prime-numbers between 12 and 150

}
