% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{primlist}
\alias{primlist}
\title{A vector containing primes between 1 and 99,999,999}
\format{
A vector containing primes between 1 and 99,999,999
}
\usage{
data(primlist)
}
\description{
Contains primes between 1 and 99,999,999
}
\details{
Variables in the dataset:
\itemize{
  \item primlist. A vector containing primes between 1 and 99,999,999
}
}
\keyword{datasets}
