\name{schoolmath-package}
\alias{schoolmath-package}
\alias{schoolmath}
\docType{package}
\title{
Functions and datasets for math used in school
}
\description{
This package contains functions and datasets for math taught in school. A main focus is set to prime-calculation
}
\details{
\tabular{ll}{
Package: \tab schoolmath\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2007-06-08\cr
License: \tab GPL version 2 or newer \cr
}
}
\author{
Joerg Schlarmann, Josef Wienand

Maintainer: Joerg Schlarmann <schlarmann@produnis.de>
}
\keyword{ package }
