% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scm}
\alias{scm}
\title{calculating the smallest common multiple of two numbers}
\usage{
scm(x, y)
}
\arguments{
\item{x}{first number}

\item{y}{second number
#'}
}
\value{
numeric least common multiple
}
\description{
calculating the smallest common multiple of two numbers
}
\examples{
scm(3528, 3780)

}
