% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratics.R
\name{solvefactorisingproblem}
\alias{solvefactorisingproblem}
\title{solvefactorisingproblem(coeffs)}
\usage{
solvefactorisingproblem(coeffs)
}
\arguments{
\item{coeffs}{vector of 3 Coefficients}
}
\value{
List containing  String Representation of Solution
}
\description{
Solve Factorising Problem
}
\examples{
x = solvefactorisingproblem(c(1,2,1))
cat(x[[1]])
}
