% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gcd}
\alias{gcd}
\title{Greatest common divisor of two numbers}
\usage{
gcd(x, y)
}
\arguments{
\item{x}{first number}

\item{y}{second number
#'}
}
\value{
numeric greatest common divisor
}
\description{
Greatest common divisor of two numbers
}
\examples{
gcd(42, 56)

}
