% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is.odd}
\alias{is.odd}
\title{checks if a number or vector is odd}
\usage{
is.odd(x)
}
\arguments{
\item{x}{the number or vector to check
#'}
}
\value{
true or false
}
\description{
checks if a number or vector is odd
}
\examples{
is.odd(45)
is.odd(46)
x <- c(1,2,3,4,5, 6, 7)
is.odd(x)

}
