% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is.prim}
\alias{is.prim}
\title{check whether a vector contains prime-numbers}
\usage{
is.prim(y)
}
\arguments{
\item{y}{the number or vector to check}
}
\value{
true or false
}
\description{
check whether a vector contains prime-numbers
}
\examples{
is.prim(8)  # this will return FALSE
is.prim(11) # this will return TRUE

x <- c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11)
is.prim(x)

}
