% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is.whole}
\alias{is.whole}
\title{check whether a vector contains numbers with decimal places}
\usage{
is.whole(x)
}
\arguments{
\item{x}{the number or vector to check
#'}
}
\value{
true or false
}
\description{
check whether a vector contains numbers with decimal places
}
\examples{
is.whole(3.12)  # this will return FALSE
is.whole(2)     # this will return TRUE

x <- c(1, 2, 3, 4, 5.5, 6.03, 23.07)
is.whole(x)

}
