% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prime.factor}
\alias{prime.factor}
\title{This function calculates the prime-factors of a number}
\usage{
prime.factor(n)
}
\arguments{
\item{n}{the number to be checked
#'}
}
\value{
a vector with the prime factors
}
\description{
This function calculates the prime-factors of a number
}
\examples{
prime.factor(21)
prime.factor(100)

}
