% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{primlist}
\alias{primlist}
\title{A vector containing primes from 0 to 9999999}
\format{
A vector containing primes from 0 to 9999999
}
\usage{
data(primlist)
}
\description{
Contains primes from 0 to 9999999
}
\details{
Variables in the dataset:
\itemize{
  \item primlist. A vector containing primes from 0 to 9999999
}
}
\keyword{datasets}
