\name{scm}
\alias{scm}
\title{ calculating the smallest common multiple of two numbers }
\description{
  This function calculates the smallest common multiple (least common multiple) of two numbers
}
\usage{
scm(x, y)
}
\arguments{
  \item{x}{ first number }
  \item{y}{ second number }
}


\author{ Joerg Schlarmann }
\seealso{ \code{\link{gcd}} }
\examples{
scm(3528, 3780)
}

\keyword{ math }

